Authorization og Tag, Hobby and User done

+++ app/Http/Controllers/TagController.php
@@ -83,6 +83,8 @@ class TagController extends Controller
       */
      public function edit(Tag $tag)
      {
+         abort_unless(Gate::allows('update' , $tag), 403,'You are not authorized.');
+
          return view('tag.edit')->with([
            'tag'=>$tag
          ]);
@@ -97,6 +99,9 @@ class TagController extends Controller
       */
      public function update(Request $request, Tag $tag)
      {
+
+         abort_unless(Gate::allows('update' , $tag), 403,'You are not authorized.');
+
        $request->validate([
          'name' => 'required|min:3',
          'style' => 'required|min:3',
@@ -120,6 +125,8 @@ class TagController extends Controller
       */
      public function destroy(Tag $tag)
      {
+         abort_unless(Gate::allows('delete' , $tag), 403,'You are not authorized.');
+
          $oldName = $tag->name;
          $tag->delete();
          return $this->index()->with(
